refactor: move workflow to repository
