+++ backend/pkg/router/router_api.go
@@ -114,7 +114,7 @@ func setApiRouter(r *resource) {
 
 	alertApi := r.mux.Group("/api/alerts")
 	{
-		alertHandler := alerts.New(r.logger, r.ch, r.pkg_db, r.k8sApi, r.prom, r.alertWorkflow)
+		alertHandler := alerts.New(r.logger, r.ch, r.pkg_db, r.k8sApi, r.prom, r.dify)
 		alertApi.POST("/event/list", alertHandler.AlertEventList())
 		alertApi.POST("/event/detail", alertHandler.AlertEventDetail())
 		alertApi.POST("/inputs/alertmanager", alertHandler.InputAlertManager())
@@ -241,7 +241,7 @@ func setApiRouter(r *resource) {
 
 	alertInputApi := r.mux.Group("/api/alertinput")
 	{
-		handler := alertinput.New(r.logger, r.ch, r.prom, r.pkg_db, r.alertWorkflow)
+		handler := alertinput.New(r.logger, r.ch, r.prom, r.pkg_db, r.dify)
 		alertInputApi.POST("/event/source", handler.SourceHandler())
 		alertInputApi.POST("/event/json", handler.JsonHandler())
 		alertInputApi.POST("/source/create", handler.CreateAlertSource())
